(GH1972) removes attempt to innappropriately test browser functionality
